requirements.txt
hat-aio ~=0.7.9
hat-doit ~=0.15.7
hat-json ~=0.5.24
